created a navbar component
