+++ client/index.html
@@ -21,6 +21,7 @@
     </script>
   </head>
   <body>
-    <div id='calendar'></div>
+    <div id="root"></div>
+    <script type="module" src="/src/main.jsx"></script>
   </body>
 </html>
