--- client/src/App.css
@@ -1,17 +1,0 @@
-/* body {
-  font-family: 'Rubik', sans-serif;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  font-family: 'Rubik Mono One', sans-serif;
-}
-
-.small {
-  font-weight: 300;
-  font-style: italic;
-} */
