+++ client/src/pages/homepage.css
@@ -1,0 +1,3 @@
+.custom-margin-top {
+  margin-top: 5rem;
+}
