+++ client/src/pages/homepage.jsx
@@ -1,0 +1,15 @@
+import './homepage.css';
+import { Container } from "react-bootstrap"
+
+
+export default function Homepage() {
+  return (
+    
+    <div className="custom-margin-top">
+        <h1>hello</h1>
+        <h1>hello</h1>
+        <h1>hello</h1>
+        <h1>hello</h1>
+    </div>
+  )
+}
